requirements.txt
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
mysqlclient==1.4.4
oauthlib==3.1.0
pycparser==2.19
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
spotipy==2.4.4
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
